KrishiDosti_AI_Models/Soilapi/requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python
pillow
